+++ logging/application_board-service_2021-09-25.log
@@ -941,7 +941,7 @@ Caused by: java.net.ConnectException: Connection timed out: connect
 2021-09-25T10:48:43.733+0900  INFO 18640 --- [nio-8080-exec-1] o.a.c.c.C.[Tomcat].[localhost].[/]       [ 173] : Initializing Spring DispatcherServlet 'dispatcherServlet'
 2021-09-25T10:48:43.735+0900  INFO 18640 --- [nio-8080-exec-1] o.s.web.servlet.DispatcherServlet        [ 525] : Initializing Servlet 'dispatcherServlet'
 2021-09-25T10:48:43.739+0900  INFO 18640 --- [nio-8080-exec-1] o.s.web.servlet.DispatcherServlet        [ 547] : Completed initialization in 2 ms
-2021-09-25T10:48:43.875+0900  WARN 18640 --- [nio-8080-exec-1] .m.m.a.ExceptionHandlerExceptionResolver [ 207] : Resolved [org.springframework.http.converter.HttpMessageNotReadableException: Required request body is missing: public org.springframework.http.ResponseEntity com.bithumb.board.controller.CommentController.createComment(com.bithumb.board.domain.Comment,long)]
+2021-09-25T10:48:43.875+0900  WARN 18640 --- [nio-8080-exec-1] .m.m.a.ExceptionHandlerExceptionResolver [ 207] : Resolved [org.springframework.http.converter.HttpMessageNotReadableException: Required request body is missing: public org.springframework.http.ResponseEntity com.bithumb.board.comment.controller.CommentController.createComment(com.bithumb.board.comment.domain.Comment,long)]
 2021-09-25T10:50:21.037+0900  WARN 18640 --- [nio-8080-exec-2] o.h.engine.jdbc.spi.SqlExceptionHelper   [ 137] : SQL Error: 1062, SQLState: 23000
 2021-09-25T10:50:21.039+0900 ERROR 18640 --- [nio-8080-exec-2] o.h.engine.jdbc.spi.SqlExceptionHelper   [ 142] : (conn=64) Duplicate entry '1' for key 'PRIMARY'
 2021-09-25T10:50:21.082+0900  WARN 18640 --- [nio-8080-exec-2] .m.m.a.ExceptionHandlerExceptionResolver [ 207] : Resolved [org.springframework.dao.DataIntegrityViolationException: could not execute statement; SQL [n/a]; constraint [PRIMARY]; nested exception is org.hibernate.exception.ConstraintViolationException: could not execute statement]
