+++ logging/spring.log
@@ -941,7 +941,7 @@ Caused by: java.net.ConnectException: Connection timed out: connect
 2021-09-25 10:48:43.733  INFO 18640 --- [http-nio-8080-exec-1] o.a.c.c.C.[Tomcat].[localhost].[/]       : Initializing Spring DispatcherServlet 'dispatcherServlet'
 2021-09-25 10:48:43.735  INFO 18640 --- [http-nio-8080-exec-1] o.s.web.servlet.DispatcherServlet        : Initializing Servlet 'dispatcherServlet'
 2021-09-25 10:48:43.739  INFO 18640 --- [http-nio-8080-exec-1] o.s.web.servlet.DispatcherServlet        : Completed initialization in 2 ms
-2021-09-25 10:48:43.875  WARN 18640 --- [http-nio-8080-exec-1] .m.m.a.ExceptionHandlerExceptionResolver : Resolved [org.springframework.http.converter.HttpMessageNotReadableException: Required request body is missing: public org.springframework.http.ResponseEntity com.bithumb.board.controller.CommentController.createComment(com.bithumb.board.domain.Comment,long)]
+2021-09-25 10:48:43.875  WARN 18640 --- [http-nio-8080-exec-1] .m.m.a.ExceptionHandlerExceptionResolver : Resolved [org.springframework.http.converter.HttpMessageNotReadableException: Required request body is missing: public org.springframework.http.ResponseEntity com.bithumb.board.comment.controller.CommentController.createComment(com.bithumb.board.comment.domain.Comment,long)]
 2021-09-25 10:50:21.037  WARN 18640 --- [http-nio-8080-exec-2] o.h.engine.jdbc.spi.SqlExceptionHelper   : SQL Error: 1062, SQLState: 23000
 2021-09-25 10:50:21.039 ERROR 18640 --- [http-nio-8080-exec-2] o.h.engine.jdbc.spi.SqlExceptionHelper   : (conn=64) Duplicate entry '1' for key 'PRIMARY'
 2021-09-25 10:50:21.082  WARN 18640 --- [http-nio-8080-exec-2] .m.m.a.ExceptionHandlerExceptionResolver : Resolved [org.springframework.dao.DataIntegrityViolationException: could not execute statement; SQL [n/a]; constraint [PRIMARY]; nested exception is org.hibernate.exception.ConstraintViolationException: could not execute statement]
