--- src/main/java/com/bithumb/board/domain/Authority.java
@@ -1,5 +1,0 @@
-package com.bithumb.board.domain;
-
-public enum Authority {
-    ROLE_USER, ROLE_ADMIN
-}
