+++ src/main/java/com/bithumb/board/reply/repository/ReplyRepository.java
@@ -1,0 +1,8 @@
+package com.bithumb.board.reply.repository;
+
+
+import com.bithumb.board.reply.domain.Reply;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ReplyRepository extends JpaRepository<Reply, Long > {
+}
