--- src/main/java/com/bithumb/board/repository/CommentRepository.java
@@ -1,16 +1,0 @@
-package com.bithumb.board.repository;
-
-import com.bithumb.board.domain.Board;
-import com.bithumb.board.domain.Comment;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-
-@Repository
-public interface CommentRepository extends JpaRepository<Comment, Long > {
-
-    Page<Comment> findCommentsByBoard(Board board, Pageable pageable);
-
-}
