--- src/main/java/com/bithumb/board/repository/ReplyRepository.java
@@ -1,8 +1,0 @@
-package com.bithumb.board.repository;
-
-
-import com.bithumb.board.domain.Reply;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ReplyRepository extends JpaRepository<Reply, Long > {
-}
