--- src/main/java/com/bithumb/board/repository/UserRepository.java
@@ -1,10 +1,0 @@
-package com.bithumb.board.repository;
-
-import com.bithumb.board.domain.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface UserRepository extends JpaRepository<User, Long > {
-
-}
