+++ src/main/java/com/bithumb/board/user/domain/Authority.java
@@ -1,0 +1,5 @@
+package com.bithumb.board.user.domain;
+
+public enum Authority {
+    ROLE_USER, ROLE_ADMIN
+}
