+++ src/main/java/com/bithumb/board/user/repository/UserRepository.java
@@ -1,0 +1,10 @@
+package com.bithumb.board.user.repository;
+
+import com.bithumb.board.user.domain.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface UserRepository extends JpaRepository<User, Long > {
+
+}
